# Reuse internal buffers