--- a/src/main/java/com/github/luben/zstd/BufferPool.java
+++ b/src/main/java/com/github/luben/zstd/BufferPool.java
@@ -1,5 +1,6 @@
 package com.github.luben.zstd;
 
+import java.lang.ref.SoftReference;
 import java.util.Map;
 import java.util.HashMap;
 import java.util.Deque;
@@ -9,26 +10,41 @@ import java.util.ArrayDeque;
  * An pool of buffers which uses a simple reference queue to recycle old buffers.
  */
 class BufferPool {
-    private static final Map<Integer, Deque<byte[]>> queues = new HashMap<Integer, Deque<byte[]>>();
+    private static final Map<Integer, SoftReference<BufferPool>> pools = new HashMap<Integer, SoftReference<BufferPool>>();
 
-    private static synchronized Deque<byte[]> getQueue(int length) {
-        Deque<byte[]> queue = queues.get(length);
-        if (queue == null) {
-            queue = new ArrayDeque<byte[]>();
-            queues.put(length, queue);
+    static BufferPool get(int length) {
+        synchronized (pools) {
+            SoftReference<BufferPool> poolReference = pools.get(length);
+            BufferPool pool;
+            if (poolReference == null || (pool = poolReference.get()) == null) {
+                pool = new BufferPool(length);
+                poolReference = new SoftReference<BufferPool>(pool);
+                pools.put(length, poolReference);
+            }
+            return pool;
         }
-        return queue;
     }
 
-    static synchronized byte[] checkOut(int length) {
-        byte[] buffer = getQueue(length).pollFirst();
+    private final int length;
+    private final Deque<byte[]> queue;
+
+    BufferPool(int length) {
+        this.length = length;
+        this.queue = new ArrayDeque<byte[]>();
+    }
+
+    synchronized byte[] checkOut() {
+        byte[] buffer = queue.pollFirst();
         if (buffer == null) {
             buffer = new byte[length];
         }
         return buffer;
     }
 
-    static synchronized void checkIn(byte[] buffer) {
-        getQueue(buffer.length).addLast(buffer);
+    synchronized void checkIn(byte[] buffer) {
+        if (length != buffer.length) {
+            throw new IllegalStateException("buffer size mismatch");
+        }
+        queue.addLast(buffer);
     }
 }