@luben 2020-10-17
Hi, sorry for the delay - I was away for a few weeks.

@chia7712 , can you take a look - this approach seams good to me - it uses internal buffer poll so it will work with the reflection. Do you see any disadvantage for your use case?